move cuda gencodes rule
